refactor: remove old configuration store
